refactor: 重构 model 层和响应体设计

+++ internal/model/message.go
@@ -3,9 +3,9 @@ package model
 import "time"
 
 type Message struct {
-	ID         uint64    `gorm:"primaryKey;autoIncrement"`
-	SenderID   uint64    `gorm:"index;not null"`
-	ReceiverID uint64    `gorm:"index;not null"`
+	ID         uint      `gorm:"primaryKey;autoIncrement"`
+	SenderID   uint      `gorm:"index;not null"`
+	ReceiverID uint      `gorm:"index;not null"`
 	Content    string    `gorm:"type:varchar(1024);not null"`
 	CreatedAt  time.Time `gorm:"index"`
 }
